pack-objects: charge a per-object floor in region weights
The unit split treats a region's byte count as its delta search cost,
but the search also pays a window of try_delta calls per object no
matter how small the objects are. A region of many tiny objects looks
light in bytes and stays below the split threshold, even though its
search time is far above the unit target. The commits are the worst
case: the path walk emits all of them as one region.

Charge every object a flat 16KB on top of its size. That figure is
measured on chromium.googlesource.com/chromium/src: a try_delta round
on a tiny object costs about 5us and blob-heavy units process about
70MB/s, so one object costs as much as roughly 16KB of blob bytes.
The padded weight feeds the split gate, the unit count and the
heaviest-first queue order, so many-object regions now split into
units and start early instead of running as one long straggler.

Packing a 40,000-commit slice of that repository with --path-walk
--window=50 on 32 threads spent 10.2s in a single thread on the
commit region (40,000 objects, 51MB), most of it after every other
unit had finished: cores drop to 1 for the last 5.5s of the pass.
With the per-object charge the region splits into 22 units, the
pass stays at ~31 busy cores, and the phase takes 6.6s instead of
12.2s (total 17.1s instead of 22.5s). The resulting pack grows by
0.07%: commits gain little from a wide shared window. At the default
--window=10 the timings and pack size are unchanged; the commit
region's search is short enough there to hide inside the parallel
phase.

+++ pack-objects.h
@@ -127,8 +127,12 @@ struct object_entry {
 struct packing_region {
 	size_t start;
 	size_t nr;
-	/* Total object bytes in the region, as a work estimate. */
-	size_t weight;
+	/*
+	 * Total object bytes in the region, as a work estimate. The
+	 * per-object cost floor overflows a 32-bit size_t already at a
+	 * quarter million objects, so keep the weight 64-bit.
+	 */
+	uintmax_t weight;
 };
 
 struct packing_data {
